requirements.txt
pipenv
mysql-connector-python